// src/utils/types.ts
export type TReviewCard = {
    avatar: string
    id: number
    image: string
    job: string
    name: string
    rating: number
    text: string
    title: string
}


export type TTipsAndTricksCard = {
    date: string
    icon: string
    id: number
    image: string
    name: string
    read: string
    text: string
    title: string
}

export type TRecommendationCard = {
    feature: string
    icon: string
    id: number
    image: string
    name: string
    avatar: string
    place: string
    price: string
    title: string

}

export type THouseStatCard = {
    avatars: []
    id: number
    header: string
    description: string

}